+++ rules/stylistic-issues-a2m.js
@@ -70,6 +70,10 @@ module.exports = {
     'error',
     'expression'
   ],
+  'function-paren-newline': [
+    'error',
+    'consistent'
+  ],
   'id-blacklist': 'off',
   'id-length': 'off',
   'id-match': [